pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ameera sherin <ameera@keyvalue.systems>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.1"
sqlalchemy = "^2.0.39"
python-dotenv = "^1.0.1"
authlib = "^1.5.1"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
